% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_emp_data.R
\docType{data}
\name{emp_data}
\alias{emp_data}
\title{Employment Data from PLFS}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43345 rows and 9 columns.
}
\source{
Periodic Labour Force Survey (PLFS) Calendar 2024
}
\usage{
emp_data
}
\description{
A dataset containing employment information from the Periodic Labour Force Survey (PLFS)
for India, including demographic and employment characteristics by state.
}
\examples{
\dontrun{
data(emp_data)
head(emp_data)
}
}
\keyword{datasets}
