% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map_of_india.R
\docType{data}
\name{map_of_india}
\alias{map_of_india}
\title{Map of India with State Boundaries}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 2 columns.
}
\source{
Derived from d3po package subnational boundaries
}
\usage{
map_of_india
}
\description{
A spatial dataset containing the geographic boundaries of Indian states and union territories.
This dataset provides the geometry for creating choropleth maps of India.
}
\examples{
\dontrun{
data(map_of_india)
plot(map_of_india["region"])
}
}
\keyword{datasets}
